api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
sqlalchemy==2.0.27
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.6.3
fal-serverless==0.8.5